Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
google-generativeai==0.3.2
numpy==1.24.3
scikit-learn==1.3.0